Added hook useObjectItems to fetch object items
